Use the same source files in Demo and update code to Swift 5
