+++ Pod/Classes/ModernSearchBarCell.swift
@@ -15,7 +15,7 @@ public class ModernSearchBarCell: UITableViewCell {
     let imgModernSearchBar = UIImageView()
     var labelModelSearchBar = UILabel()
     
-    override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         self.setup()
     }
